fix delete products and menus when dont have photos

+++ src/Application/Orchestrator/ProductOrchestrator.php
@@ -241,7 +241,9 @@ final class ProductOrchestrator extends AbstractController
             );
             return;
         }
-        $this->managePhoto->deletePhoto($product->getPhotos()->first()->getPhotoPath());
+        if($product->getPhotos()->first()){
+            $this->managePhoto->deletePhoto($product->getPhotos()->first()->getPhotoPath());
+        }
         $this->productRepository->remove($product, true);
         
         $this->addFlash(
